Swap 'DailyRewardTest' and 'DailyRewardTest2' / Change excute test `Theory` to `Fact `
