+++ .Lib9c.Tests/Action/DailyRewardTest2.cs
@@ -1,12 +1,10 @@
 ﻿namespace Lib9c.Tests.Action
 {
-    using System.Linq;
     using Libplanet;
     using Libplanet.Action;
     using Libplanet.Crypto;
     using Nekoyume;
     using Nekoyume.Action;
-    using Nekoyume.Model.Mail;
     using Nekoyume.Model.State;
     using Serilog;
     using Xunit;
@@ -57,11 +55,10 @@
                 .SetState(_avatarAddress, avatarState.Serialize());
         }
 
-        [Theory]
-        [InlineData(400000, 10)]
-        public void Execute(int itemId, int itemCount)
+        [Fact]
+        public void Execute()
         {
-            var dailyRewardAction = new DailyReward
+            var dailyRewardAction = new DailyReward2
             {
                 avatarAddress = _avatarAddress,
             };
@@ -69,7 +66,6 @@
             {
                 BlockIndex = 0,
                 PreviousStates = _initialState,
-                Random = new TestRandom(),
                 Rehearsal = false,
                 Signer = _agentAddress,
             });
@@ -77,22 +73,12 @@
             var gameConfigState = nextState.GetGameConfigState();
             var nextAvatarState = nextState.GetAvatarState(_avatarAddress);
             Assert.Equal(gameConfigState.ActionPointMax, nextAvatarState.actionPoint);
-            Assert.Single(nextAvatarState.mailBox);
-            var mail = nextAvatarState.mailBox.First();
-            var rewardMail = mail as DailyRewardMail;
-            Assert.NotNull(rewardMail);
-            var rewardResult = rewardMail.attachment as DailyReward.DailyRewardResult;
-            Assert.NotNull(rewardResult);
-            Assert.Single(rewardResult.materials);
-            var material = rewardResult.materials.First();
-            Assert.Equal(itemId, material.Key.Id);
-            Assert.Equal(itemCount, material.Value);
         }
 
         [Fact]
         public void ExecuteThrowFailedLoadStateException()
         {
-            var action = new DailyReward
+            var action = new DailyReward2
             {
                 avatarAddress = _avatarAddress,
             };
